requirements.txt
arxiv==2.1.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sentence-transformers
transformers
chromadb==0.5.3
rank-bm25==0.2.2
numpy==1.26.4
pandas==2.2.2
fastapi==0.111.0
uvicorn==0.30.1
torch